Refactor SGQuery creation to keep baseURL internal
This creates a simpler SGQuery constructor that only needs the URL path and not the API's baseUrl. For these queries, the baseURL is used to construct the full URL only at the time the request is created. This allows the baseURL to be changed (perhaps between debug URLs) without requiring the SGQuery objects be recreated. URLs on other hosts can still be queried as before, but using a different constructor.

Also note that paths should have a leading "/" in most cases. This makes it easier to search the codebase for a specific endpoint, i.e. "/events" returns many fewer results than "event".