Move recursive deep fetch from SGDataManager to SGItem
This method was not specific to the data manager in question, and I needed to expose it for use outside of the data manage in our app. Note that simply overriding fetch to also fetch the children would not work, becuase it is a common pattern for subclasses of SGItem to override fetch without calling super.